add repo files
A PHP based javascript minification tool, based on Google's closure
compiler. You cna run this locally in WAMP, or MAMP. Choose the file
you want to be compressed, choose the compression level, and the name
of the file you would like output and this tool will pass your script
to the closure compiler tool, get back the returned content, and save
it in a file for you.

There is a constant called SCRIPT_PATH defined in the php file, change
this to your desired location for file destination. all else should
work out of the box.